refactor(api): global ai service
